Refactored some of the message code to make debugging easier.
There's now a message type and is used for both sending and receiving.  The receive code is only partway there (it half reads from the parsed version and half from the bytes read from websockets), but it's able to using the same debugging code both ways, which is a step up.

Signed-off-by: Scott Laird <scott@sigkill.org>