Switched from staticly defined displays to dynamically defined displays.
Not complete yet, but the Display type went from a 16-bit int that just held the display ID for the Loupedeck protocol to a struct that actually describes the display.  Hooks are in place for this to be set up dynamically at runtime, but that code isn't present yet.  This changes the API quite a bit; some things moved from Loupedeck.<> to Display.<> (Height, Draw, maybe others).

This is still a work in progress, but it works with my Loupedeck Live.  It still doesn't work with my Loupedeck CT.

Signed-off-by: Scott Laird <scott@sigkill.org>